ec-payment/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.12.3,<3.0.0
midtransclient>=1.4.2,<2.0.0
python-dotenv>=1.2.1,<2.0.0
